Huge update. New object for showing articles in the application, DB connection object instead of class DatabaseCommunication, rename of AgencyExtended, inclusion of Address object for future use, test class for database stuff with several test ready-to-go!
DBobject initiates a connection when it is created as well as
registering the driver. After that it uses preparedStatements to make
queries and batch updates are used for inserts/updates/deletions.

ShownObject is supposed to be a general object for showing information
in the application. It has 15 simplestringproperties and constructors
for scenarios where an object with 1 up to 15 variables is created. The
idea is to create generic routines for saving query information.

AgencyExtended renamed AgencyExt for simplicity purposes.

Address object added "just in case" we go away from the ShownObject
option and need to retrieve ONLY address information, however unlikely.

DBtests contain several ready tests if you need a look at the code.