Found issues in database object
We needed to always close connections/result sets and so on when
selecting prior to an insertion, otherwise we got a "locked" error.
Fixed this by making setConnection private and starting the connection
with each execute method. YOU WILL ALWAYS NEED TO CLOSE CONNECTIONS
MANUALLY AFTER YOU'VE RETRIEVED/UPDATED INFORATION. This uses the same
method as before, closeConnection();