Automatic adding a default rating to newly created agencies (0, N/A) and added logo support. You can now update pictures on ads.
Added support for updating pictures in ads via the memberPage. Also
added a default rating when agencies are created, which is needed for
them to be able to log in. Try catch added to main to detect errors with
automatic creation of ratings when logging in after creating an account.